Add a specLevel property to MeasurementBase
Each Measurement has a specLevel property that can be set to a
specification level name (design, minimum, stretch) to declare that this
measurement is specific to that level. This happens when a measurement
is dependent on a specification level of another metric. This property
will be None if the Measurement is not level-specific. Serialized to
JSON as spec_level.