Add Job class to base module
Job takes role of JobSerializer. It is now able to incrementally  build
lists of measurements and blobs to serialize since their respective
unique IDs are now maintained as sets and tracked.

Also adds unique IDs to the MeasurementBase and BlobBase-derived objects
so that they can be uniquely identified. For example, the list of blobs
serialized to JSON for each Measurement object consists of their IDs,
referencing the Blobs JSON document.